cleaning and documenting tcp/ip communication
