+++ documentation/usage/api/utility_api/tcp_ip.rst
@@ -1,6 +1,8 @@
+.. _tcpip_API:
+
 TCP/IP related methods
 **********************
 
-.. automodule:: pymodaq.daq_utils.daq_utils
-   :members: send_string, send_scalar, get_string, get_int, get_scalar, get_list, send_list, get_array, send_array, message_to_bytes
-             check_sended, check_received_length
+.. automodule:: pymodaq.daq_utils.tcpip_utils
+   :members:
+
